Create noitasaver.py
Hi, 

I made this save scum program mostly as an exercise for myself, but it should be useable for everyone if you follow these steps (it's also explained in the code itself):
1. 
   find where you game files are, specifically where your save00 folder is 
2. 
   find a place you want to save your gamefiles in 
3. 
   for convenience sake, replace the note after destination with the file path you want to save your saves in 
   IMPORTANT: make sure you add a second ' \ ' after each ' \ ' in the file paths, otherwise python won't read it correctly 
4.  
   on line 5, replace the note (so everything after the = mark) with your file path to the repository save00 in your game files.
   on line 15, 21 and 27 replace the notes with your destination but increment each file with 1. so for the first save spot it should end with save01, the second spot 
   save02 and the third spot save03
5. 
   do the previous step but now on lines 38, 44 and 50. 

if everything is done correctly, you should be able to run python code which allows you to save your games. Use it at your own risk though, if you mess up the file paths to where you want to save your games it could lead to the replacement of that folder with the save00 folder. So if you really want to be careful make another folder where in you want to save it. If you have any feedback or questions you can reach out to me on discord:  Willie#8572